requirements.txt
amqp==5.3.1
asgiref==3.7.2
billiard==4.2.1
celery==5.4.0
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.3
defusedxml==0.7.1
Django==4.2.4
django-admin-notification==0.2
django-filter==23.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.7
generics==6.0.0
idna==3.4
inflection==0.5.1
kombu==5.4.2
oauthlib==3.2.2
packaging==23.1
prompt_toolkit==3.0.50
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
stripe==11.5.0
typing_extensions==4.12.2
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.13
